+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Upcoming
+* [fix] notification: use event.waitUntil to fix bkg processing notification issue 
+
 ## v0.16.16 - 2017-05-30
 * [fix] notifications: step 2: functions: 
     * ignore push for connected clients.